+++ memory-viz/src/automate.ts
@@ -73,6 +73,7 @@ function drawAutomated(
         roughjs_config: configuration.roughjs_config,
         global_style: configuration.global_style,
         theme: configuration.theme,
+        interactive: configuration.interactive,
     });
 
     m.drawAll(StackFrames);