+++ memory-viz/src/style.ts
@@ -147,6 +147,8 @@ const DARK_THEME_CSS = `
             --fade-box-line-color: #666666;
         
             --hide-box-fill: #121212;
+
+            --highlight-object-fill: rgba(72, 207, 173, 0.3);
         }
         [data-theme="dark"] text.default,
         [data-theme="dark"] text.variable {
@@ -180,6 +182,8 @@ const HIGH_CONTRAST_THEME_CSS = `
             --fade-box-line-color: #FFFFFF;
             
             --hide-box-fill: #000000;
+
+            --highlight-object-fill: rgba(0, 0, 255, 0.5);
         }
         [data-theme="high-contrast"] text.default,
         [data-theme="high-contrast"] text.variable {
@@ -207,6 +211,8 @@ function setStyleSheet(
     global_style?: string,
     theme?: string
 ) {
+    const interactiveCursor = memory_model.interactive ? "pointer" : "auto";
+
     const styles = `
         :root {               
         --fade-text-color: ${config.text_color};
@@ -226,6 +232,8 @@ function setStyleSheet(
         --primitive-value-color: ${config.value_color};
         --id-text-color: ${config.id_color};
         --default-font-size: ${config.font_size}px;
+
+        --highlight-object-fill: rgba(255, 255, 0, 0.6);
     } 
     
         text {
@@ -245,6 +253,7 @@ function setStyleSheet(
             text-anchor: start;
         }
         text.id { 
+            cursor: ${interactiveCursor};
             fill: ${config.id_color};
             text-anchor: middle;
         }
@@ -258,6 +267,9 @@ function setStyleSheet(
         }
         path {
             stroke: ${config.rect_style?.stroke ?? "rgb(0,0,0)"};
+        }
+        .highlighted path {
+            fill: var(--highlight-object-fill) !important;
         }
     `;